+++ benefits/WingspanBenefits/BenefitsService.cs
@@ -51,10 +51,10 @@ namespace WingspanBenefits
     {
         public SDKConfig SDKConfiguration { get; private set; }
         private const string _language = "csharp";
-        private const string _sdkVersion = "2.0.2";
-        private const string _sdkGenVersion = "2.209.0";
+        private const string _sdkVersion = "2.0.3";
+        private const string _sdkGenVersion = "2.210.6";
         private const string _openapiDocVersion = "1.0.0";
-        private const string _userAgent = "speakeasy-sdk/csharp 2.0.2 2.209.0 1.0.0 WingspanBenefits";
+        private const string _userAgent = "speakeasy-sdk/csharp 2.0.3 2.210.6 1.0.0 WingspanBenefits";
         private string _serverUrl = "";
         private ISpeakeasyHttpClient _defaultClient;
         private ISpeakeasyHttpClient _securityClient;